cryptx/ai-service/main.py
from fastapi import FastAPI, HTTPException
from fastapi.middleware.cors import CORSMiddleware
from pydantic import BaseModel
from typing import List, Optional
import os
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

# Initialize FastAPI app
app = FastAPI(
    title="Crypto Portfolio AI Service",
    description="AI-powered analytics for cryptocurrency portfolios",
    version="1.0.0"
)

# CORS configuration
app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],  # Configure based on your needs
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

# Request/Response Models
class TokenData(BaseModel):
    symbol: str
    balance: float
    usdValue: float
    percentage: float

class PortfolioData(BaseModel):
    walletAddress: str
    chainId: int
    totalValue: float
    tokens: List[TokenData]

class ForecastRequest(BaseModel):
    tokens: List[str]
    forecastDays: List[int] = [7, 30]

class DiversificationRequest(BaseModel):
    walletAddress: str
    portfolioData: PortfolioData

class RiskAnalysisRequest(BaseModel):
    walletAddress: str
    portfolioData: PortfolioData

# Health check endpoint
@app.get("/health")
async def health_check():
    return {
        "status": "ok",
        "service": "AI Service",
        "version": "1.0.0"
    }

# API info endpoint
@app.get("/")
async def root():
    return {
        "message": "Crypto Portfolio AI Service",
        "version": "1.0.0",
        "endpoints": {
            "health": "/health",
            "forecast": "/api/forecast-prices",
            "diversification": "/api/analyze-diversification",
            "risk": "/api/analyze-risk"
        }
    }

# Placeholder endpoints (to be implemented)
@app.post("/api/forecast-prices")
async def forecast_prices(request: ForecastRequest):
    """
    Generate price forecasts using LSTM model
    """
    # TODO: Implement LSTM price forecasting
    return {
        "message": "Price forecasting endpoint - to be implemented",
        "tokens": request.tokens,
        "forecastDays": request.forecastDays
    }

@app.post("/api/analyze-diversification")
async def analyze_diversification(request: DiversificationRequest):
    """
    Analyze portfolio diversification using MPT
    """
    # TODO: Implement MPT portfolio optimization
    return {
        "message": "Diversification analysis endpoint - to be implemented",
        "walletAddress": request.walletAddress,
        "totalValue": request.portfolioData.totalValue
    }

@app.post("/api/analyze-risk")
async def analyze_risk(request: RiskAnalysisRequest):
    """
    Analyze portfolio risk metrics
    """
    # TODO: Implement risk analysis
    return {
        "message": "Risk analysis endpoint - to be implemented",
        "walletAddress": request.walletAddress,
        "totalValue": request.portfolioData.totalValue
    }

if __name__ == "__main__":
    import uvicorn
    port = int(os.getenv("PORT", 8000))
    uvicorn.run(app, host="0.0.0.0", port=port)
